build-requirements.txt
python-dateutil==2.8.2
requests==2.31.0
unittest-xml-reporting==3.2.0
xunitparser==1.3.4
toposort==1.10
python-dotenv==1.0.0